feat(passquito-client-js): add Passquito client package for JS
Introduces a new package `@codemonger-io/passquito-client-js` which
facilitates to integrate Passquito into a web app. Most of the code is
extracted from the `app/src/utils` folder, and the scripts in
`passquito-client-js` folder bundle the code as a package.

The `app` package now uses `@codemonger-io/passquito-client-js`.
Includes the `app` package in the pnpm workspace.
