+++ app/src/stores/credential.ts
@@ -2,8 +2,9 @@ import { StorageSerializers, useStorage } from '@vueuse/core';
 import { defineStore } from 'pinia';
 import { computed, ref } from 'vue';
 
-import { isCognitoTokens, isPublicKeyInfo } from '../utils/passquito-types';
-import type { CognitoTokens, PublicKeyInfo } from '../utils/passquito-types';
+import { isCognitoTokens, isPublicKeyInfo } from '@codemonger-io/passquito-client-js';
+import type { CognitoTokens, PublicKeyInfo } from '@codemonger-io/passquito-client-js';
+
 import { makeValidatingSerializer } from '../utils/serializer';
 
 // possible states of the credential.
