+++ app/src/utils/errors.ts
@@ -1,0 +1,23 @@
+/**
+ * Returns if a given error object is an `AbortError`.
+ *
+ * @beta
+ */
+export function isAbortError(err: unknown): boolean {
+  if (err == null || (typeof err !== 'object' && typeof err !== 'function')) {
+    return false;
+  }
+  return (err as { name: string }).name === 'AbortError';
+}
+
+/**
+ * Returns the name of a given error object.
+ *
+ * @beta
+ */
+export function getErrorName(err: unknown): string | undefined {
+  if (err == null || (typeof err !== 'object' && typeof err !== 'function')) {
+    return undefined;
+  }
+  return (err as { name: string }).name;
+}
