+++ app/src/views/SignIn.vue
@@ -7,7 +7,7 @@ import { useWebauthn } from '../composables/webauthn';
 import { useCredentialStore } from '../stores/credential';
 import { usePasskeyCapabilityStore } from '../stores/passkey-capability';
 import { usePassquitoClientStore } from '../stores/passquito-client';
-import { getErrorName } from '../utils/passquito';
+import { getErrorName } from '../utils/errors';
 
 // router
 //
