+++ passquito-client-js/api/passquito-client-js.api.md
@@ -1,0 +1,119 @@
+## API Report File for "@codemonger-io/passquito-client-js"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+// @beta
+export interface AuthenticationSession {
+    credentialRequestOptions: CredentialRequestOptions;
+    sessionId: string;
+}
+
+// @beta
+export function checkPasskeyAuthenticationSupported(): Promise<boolean>;
+
+// @beta
+export function checkPasskeyRegistrationSupported(): Promise<boolean>;
+
+// @beta
+export interface CognitoTokens {
+    accessToken: string;
+    activatedAt: number;
+    expiresIn: number;
+    idToken: string;
+    refreshToken: string;
+}
+
+// @beta
+export interface Credentials {
+    // (undocumented)
+    publicKeyInfo: PublicKeyInfo;
+    // (undocumented)
+    tokens: CognitoTokens;
+}
+
+// @beta
+export interface CredentialsApi {
+    finishAuthentication(sessionId: string, userId: string, credential: PublicKeyCredential): Promise<CognitoTokens>;
+    finishRegistration(sessionId: string, credential: PublicKeyCredential): Promise<void>;
+    getDiscoverableCredentialRequestOptions(): Promise<CredentialRequestOptions>;
+    refreshTokens(refreshToken: string): Promise<CognitoTokens | undefined>;
+    startAuthentication(userId: string): Promise<AuthenticationSession>;
+    startRegistration(userInfo: UserInfo): Promise<RegistrationSession>;
+    startRegistrationForVerifiedUser(userInfo: VerifiedUserInfo): Promise<RegistrationSession>;
+}
+
+// @beta
+export class CredentialsApiImpl implements CredentialsApi {
+    constructor(baseUrl: string);
+    readonly baseUrl: string;
+    // (undocumented)
+    finishAuthentication(sessionId: string, userId: string, credential: PublicKeyCredential): Promise<CognitoTokens>;
+    // (undocumented)
+    finishRegistration(sessionId: string, credential: PublicKeyCredential): Promise<void>;
+    // (undocumented)
+    getDiscoverableCredentialRequestOptions(): Promise<CredentialRequestOptions>;
+    // (undocumented)
+    refreshTokens(refreshToken: string): Promise<CognitoTokens | undefined>;
+    // (undocumented)
+    startAuthentication(userId: string): Promise<{
+        sessionId: any;
+        credentialRequestOptions: CredentialRequestOptions;
+    }>;
+    // (undocumented)
+    startRegistration(userInfo: UserInfo): Promise<{
+        sessionId: any;
+        credentialCreationOptions: CredentialCreationOptions;
+    }>;
+    // (undocumented)
+    startRegistrationForVerifiedUser(userInfo: VerifiedUserInfo): Promise<any>;
+}
+
+// @beta
+export function isCognitoTokens(value: unknown): value is CognitoTokens;
+
+// @beta
+export function isPublicKeyInfo(value: unknown): value is PublicKeyInfo;
+
+// @beta
+export class PassquitoClient {
+    constructor(credentialsApi: CredentialsApi);
+    doAuthenticationCeremony(): {
+        abort: () => void;
+        credentials: Promise<Credentials>;
+    };
+    doAuthenticationCeremonyForUser(userId: string): {
+        abort: () => void;
+        credentials: Promise<Credentials>;
+    };
+    doRegistrationCeremony(userInfo: UserInfo): Promise<void>;
+    doRegistrationCeremonyForVerifiedUser(userInfo: VerifiedUserInfo): Promise<void>;
+}
+
+// @beta
+export interface PublicKeyInfo {
+    authenticatorAttachment?: string | null;
+    id: string;
+    userHandle?: string | null;
+}
+
+// @beta
+export interface RegistrationSession {
+    credentialCreationOptions: CredentialCreationOptions;
+    sessionId: string;
+}
+
+// @beta
+export interface UserInfo {
+    displayName: string;
+    username: string;
+}
+
+// @beta
+export interface VerifiedUserInfo {
+    idToken: string;
+    userInfo: UserInfo;
+}
+
+```
