+++ passquito-client-js/src/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Passquito client for browsers.
+ *
+ * @packageDocumentation
+ */
+
+export * from './credentials-api';
+export * from './credentials-api-impl';
+export * from './passkey';
+export * from './passquito-client';
+export * from './types';
