+++ passquito-client-js/src/passkey.ts
@@ -4,8 +4,8 @@
  * @remarks
  *
  * References:
- * - <https://web.dev/articles/passkey-registration>
- * - <https://www.w3.org/TR/webauthn-3/>
+ * - https://web.dev/articles/passkey-registration
+ * - https://www.w3.org/TR/webauthn-3/
  *
  * @beta
  */
@@ -48,8 +48,8 @@ export async function checkPasskeyRegistrationSupported(): Promise<boolean> {
  * @remarks
  *
  * References:
- * - <https://web.dev/articles/passkey-form-autofill>
- * - <https://www.w3.org/TR/webauthn-3/>
+ * - https://web.dev/articles/passkey-form-autofill
+ * - https://www.w3.org/TR/webauthn-3/
  *
  * @beta
  */
@@ -73,28 +73,4 @@ export async function checkPasskeyAuthenticationSupported(): Promise<boolean> {
     return false;
   }
   return true;
-}
-
-/**
- * Returns if a given error object is an `AbortError`.
- *
- * @beta
- */
-export function isAbortError(err: unknown): boolean {
-  if (err == null || (typeof err !== 'object' && typeof err !== 'function')) {
-    return false;
-  }
-  return (err as { name: string }).name === 'AbortError';
-}
-
-/**
- * Returns the name of a given error object.
- *
- * @beta
- */
-export function getErrorName(err: unknown): string | undefined {
-  if (err == null || (typeof err !== 'object' && typeof err !== 'function')) {
-    return undefined;
-  }
-  return (err as { name: string }).name;
 }
